% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp.R
\name{runListingApp}
\alias{runListingApp}
\title{Start the Survey Solutions Listing Application}
\usage{
runListingApp()
}
\description{
A wrappter function to start the application. Please make sure you have read the
documentation on how to use the app under \url{https://datanalytics.worldbank.org/SpatialSamplingManual/}
}
\details{
This application is part of the large set of tools, to facilitate survey implementation with
\href{https://docs.mysurvey.solutions/}{Survey Solutions}. Enumeration Areas, sampled for example
with the JDC-Survey Solutions Spatial Sampling application, can be used to enumerate all buildings
visible within the boundaries on a google map. The enumerate buildings can then be used for the
second stage sampling frame and to draw the survey units within the cluster from it.
}
