% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runappserver.R
\name{runListingAppServer}
\alias{runListingAppServer}
\title{Start the Survey Solutions Listing Application on Shiny Server}
\usage{
runListingAppServer()
}
\description{
A wrappter function to start the application on the SERVER. Please make sure you have read the
documentation on how to use the app under \url{https://datanalytics.worldbank.org/SpatialSamplingManual/}
}
\details{
This function only works on shiny server to do so, create a script in the app directory and
run this function. An example myListingApp.R file could be:

\if{html}{\out{<div class="sourceCode">}}\preformatted{susolisting::runListingAppServer()

}\if{html}{\out{</div>}}
}
