% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susolisting-package.R
\docType{package}
\name{susolisting-package}
\alias{susolisting}
\alias{susolisting-package}
\title{susolisting: Shiny Application for EA Listing on Satellite Imagery}
\description{
A simple shiny application to mark buildings on a google map within user provided enumeration area (EA) boundaries. The application requires the input of boundaries in ESRI shape file format, and returns a csv file with all the listed buildings. It is part of the larger series of support tools for using spatial resources with Survey Solutions.
}
\author{
\strong{Maintainer}: Michael Wild \email{mwild@worldbank.org}

}
\keyword{internal}
